% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_WGCNA_module_volcano_2var.R
\name{plot_WGCNA_module_volcano_2var}
\alias{plot_WGCNA_module_volcano_2var}
\title{Generate a volcano plot of genes from WGCNA modules and a differential expression (limma) analysis}
\usage{
\code{
plot_WGCNA_module_volcano_2var(
     topGenes, gene_modules,
     file_prefix=NULL, plotdims=c(9,9),
     kME_cut=0.75, fc_cut=log2(1.5), p_cut=0.01,
     x_lim=NULL, y_lim=NULL,
     gene_labs=FALSE, x_cut=0, y_cut=0,
     point_order="random"
     )}
}
\arguments{
\item{topGenes}{a data frame, typically the output of a call to \code{topTable}. Must contain genes, log2 fold-change, and adjusted p-values.}

\item{gene_modules}{a data frame containing the gene module memberships. Must contain, at minimum, columns for \code{gene}, \code{color_assigned}, and \code{kME_color_assigned}. If provided, the function outputs a pdf of the plot, named "{file_prefix}.pdf".}

\item{plotdims}{a numeric vector, the size (in inches) of the plotting object. Either the size of the pdf, or the size of the plotting window.}

\item{fc_cut}{numeric, the (absolute value) log2 fold-change threshold for determining significance of genes. This value is also plotted as vertical dotted lines.}

\item{p_cut}{numeric, the p-value threshold for determining significance of genes. This value is also plotted as a horizontal dotted line.}

\item{x_lim, y_lim}{numeric vectors, the lower and upper limits of the plotting space along the x- and y-axes. Passed to \code{ggplot2::xlim}.}

\item{gene_labs}{logical, whether to include gene labels for genes with extreme logFC and p-value. If \code{TRUE}, genes with values outside the labeling ellipse will be labeled.}

\item{x_cut, y_cut}{numeric, the radii of the labeling ellipse along the x- and y-axes. Genes with values outside the ellipse are labeled with gene names. Default to 0, which results in all genes being labeled.}

\item{point_order}{character string, specifying how to order the points. Currently accepted values are "random", which randomizes the order of the points, and "input", which send the points to ggplot as they are in the input data frame. Defaults to "random".}
}
\description{
Generate a volcano plot of genes from a differential expression (limma) analysis, with point
inclusion and color based on membership in modules from a WGCNA analysis. This plot can be output
to a plotting window, or to a pdf. Points can be selected for inclusion based on their connectivity
to the module eigengene (kME). Points can be labeled with gene names, and the points to be labeled
can be set based on an ellipse oriented to the x- and y-axes.
}
\details{
Genes with \code{kME_color_assigned > kME_cut} are plotted. Colors of points are determined by the values of \code{color_assigned}; if different colors are preferred, these values must be modified prior to calling \code{plot_WGCNA_module_volcano_2var}.
}
